Fix script linting during pre-commit hook
As discussed in a github issue, eslint exit code is 0 in case of
warnings; and in our linter settings, unused variables should raise
warning rather than errors. That meant, however, that lint-staged
would not fail if eslint issued warnings.

Using the max-warnings option eslint, we make lint-staged fail if
eslint encounters a warning.