Revert to using scheduler groups from timers (performance fix) (#209)
An earlier commit changed the means by which callbacks are registered to the scheduler. This was anticipated to have negligible effect, but was otherwise poorly motivated. As a result, the change had already been reverted on the active dev branch when it was later found that the change had a significant (well-characterised but poorly-understood) performance impact. Unfortunately, between the change and its reversion, the code got pushed to dev and live. The active dev branch for this code is not yet ready for merge, and the effect of this change being live  is noticeable. So this bugfix commit pushes the reversion (only) to dev.