Fix canvas overflow (#801)
The genome browser canvas, temporarily, is allowed to be taller than the container element
that we are passing to the genome browser to render into. This leads to the bottom of the canvas
being cropped out. The decision was to remove the "overflow: hidden" CSS rule
from that container element.