Add a simple table shared component (#803)
This commit introduces the Table component, which is only responsible
for styling the `table` element and its children.
There is also a RowFooter component responsible for adding a table row
with a single table cell spanning all columns of the table. When shown, it produces an effect
as if the previous table row has been expanded.