Use React 18's own apis for server-side rendering (#878)
Using React's `renderToPipeableStream` and `lazy` apis for server-side rendering and code splitting.

Removed dependencies on:
- @loadable/component, @loadable/server, and related @types packages
- react-helmet-async

Added dependency on webpack-manifest-plugin to be able to read webpack manifest on the server,
and inject it into the html.

Note: Hot module replacement has been disabled. It was already semi-broken after we added server-side rendering.