Use new variant id schema (#1004)
As agreed with the Variation team, the ids in the url that are able to uniquely identify a variant need to contain
region name, variant start coordinate, and variant name (e.g. an rsID).

Thus, we are going to use the following format:

<region_name>:<start>:<variant_name>

The variant string is read from the url, and parsed into significant parts, which are used
to direct the genome browser to bring the variant into view.
The viewport is calculated as start±25bp.
This will allow us to retire the jump file for variants.