Refactor feature sequence downloading, and add downloading of exons (#1096)
REASON FOR REFACTORING:
Prior to this commit, there used to be a function, in a worker, that combined the fetching of a single sequence
with the generation of a FASTA string for that sequence.
The problem with that approach was that if, for example, the user chose to download genomic sequences
for all transcripts in a gene, we would send multiple requests to refget, one per transcript,
which for some genes may mean hundreds of unnecessary requests.
Moreover, now that we are adding the capability of downloading exons, the number of network requests
for all exons in a gene would get much greater. All these requests are completely unnecessary,
because it is sufficient to just fetch the genomic sequence of a gene, and then to subslice transcripts and exons
out of it in the client.

SUMMARY OF THE APPROACH:
- The code for fetching the sequences and generating the FASTA strings still runs in a web worker
   (it probably doesn't have to; but there shouldn't be any harm in doing so).
- The logic for how to deal with different kinds of sequences has been moved from the main thread
   (files that used to be called fetchForGene, fetchForTranscript, etc.) into the worker.
   Now, the functions responsible for generating FASTA strings, also contain the logic
   about how these strings should be built.

NEW FUNCTIONALITY:
It is now possible to download exons, either for a single transcript, or for all transcripts in a gene.