Add notifications support to CommunicationPanel (#1163)
Up until now, the CommunicationPanel (the panel that slides from the right upon click on the conversation icon)
only showed the "Contact Us" form. This PR adds the capability to show notifications ("Messages").

The proper, long-term architecture for notifications is unclear.
For now, the requirement is only to show one notification ("welcome to beta, etc."),
which automatically opens the communication panel if the user hasn't seen it.
If the user simply closes the panel or clicks on the "use beta" button, the notification remains accessible
(shows up as a green dot next to the conversation icon). If the user chooses to "clear the message",
it disappears completely.