Add fix for IndexedDB
We've been seeing the problem that IndexedDB becomes inaccessible
right after a db version bump.

The problem is almost certainly caused by a concurrent instance
of the app running in another browser tab, and being connected to IndexedDB,
which will prevent IndexedDB from upgrading, and in fact will leave the running code
in a broken state.

This commit adds logic for handling scenarios when a tab with an older db version
is blocking another tab from db upgrade.