Fix species lozenge order with a timestamp (#1216)
This commit addresses the problem of the wrong order of species lozenges after a page refresh.

Desired behaviour (as specified by designer):
- if a user adds a single genome, it will appear to the left of the lozenges
- if a user selects multiple from one species (ie x6 human), then these should be added as a block to the left of existing lozenges
- a block of genomes would be added as they appeared in the species selector list — so the highest in the list would be left most in the added block; the lowest, the rightmost

This commit adds a timestamp to the genomes (the `addedAt` field) before saving them to the db.
When multiple genomes are selected at once, their array is first reversed before the timestamps are added;
this is to pretend that the topmost species search match was selected the latest (in order to match the design requirement).

Other changes:
- IndexedDb version is bumped from 7 to 8; and a corresponding database migration has been created.
- Some refactoring was done to the earlier db migration code, to extract the code that, due to its asynchronous nature,
   cannot run within a db version change transaction.