Update robots.txt
Add /species-selector/search/ to disallow rules for all bots.
Motivation: we are seeing a lot of requests with garbage queries
sent to the /species-selector/search route by what looks to be bingbot.

Examples of such queries are:
/species-selector/search?query=gmt+-4%3A00
/species-selector/search?query=The+KNAT3a1-WND2A%2F3A+Module+Positively+Regulates+Fiber+Secondary+Cell+Wall+Biosynthesis+in
/species-selector/search?query=gmssl%E5%AE%98%E7%BD%91%E8%AE%BF%E9%97%AE%E4%B8%8D%E4%BA%86

(The query in this last example is, according to Gemini:
URL-encoded phrase that translates from Chinese to "gmssl official website cannot be accessed")

It is unclear whether the bot that visits these urls is truly Bingbot,
as it introduces in the user-agent string:
"Mozilla/5.0 AppleWebKit/537.36 (KHTML, like Gecko; compatible; bingbot/2.0; +http://www.bing.com/bingbot.htm) Chrome/116.0.1938.76 Safari/537.36"
or it is some other bot pretending to be Bingbot.

Whichever this might be, /species-selector/search is useless for robots,
because it doesn't have any links for them to follow. So they might as well
just be forbidden from visiting this route.