Update genome selection for structural variants (#1245)
The selection of the genomes and of the location on the reference genome is reflected in the url
in the following search parameters:
- `ref-genome-id` for reference genome id
- `alt-genome-id` for comparison genome id
- `ref-location` for location in the reference genome

The page can read the search parameters from the url and update redux, which should be sufficient
to bootstrap the page. The page also fills in the genomes/location selector with the values
matching the url parameters

Also, added the "Go" button.

The logic for the StructuralVariantsTopBar component has been moved to a dedicated hook (useTopBarState).