Fix initialization of the genome browser page (#1248)
The initialization of the genome browser page is rather confusing. There are multiple useEffect hooks at play,
which modern react guidelines strongly frown upon. There is the `useGenomeBrowserTracks` hook, whose purpose is
to fetch the tracks, and to initialize their settings (i.e. check if some track settings have been saved to browser
storage, etc.). There is also the `useGenomicTracks` hook responsible for non-focus tracks, and the `useFocusTrack`
hook, responsible for the focus track.

The mistake in the code was that I must have forgotten that it's the responsibility of `useGenomeBrowserTracks` to
initialise track settings, and did so for the variant focus track in the `useFocusTrack` hook. As a result, this prevented
initialization of other track settings because of the condition here (focus track already had its settings set).